+++ daemon/builder/remotecontext/archive.go
@@ -6,9 +6,9 @@ import (
 	"path/filepath"
 
 	"github.com/docker/docker/daemon/builder"
+	"github.com/docker/docker/daemon/builder/remotecontext/internal/tarsum"
 	"github.com/docker/docker/pkg/longpath"
 	"github.com/docker/docker/pkg/system"
-	"github.com/docker/docker/pkg/tarsum"
 	"github.com/moby/go-archive/chrootarchive"
 	"github.com/moby/go-archive/compression"
 	"github.com/moby/sys/symlink"
