+++ daemon/builder/remotecontext/internal/tarsum/writercloser.go
